M05

+++ test/SphereXEngine.t.sol
@@ -25,7 +25,7 @@ contract SphereXEngineTest is Test, CFUtils {
         spherex_engine.addAllowedSender(allowed_senders);
     }
 
-    function sendNumberToEngine(int16 num) private {
+    function sendNumberToEngine(int256 num) private {
         if (num > 0) {
             spherex_engine.sphereXValidateInternalPre(num);
         } else {
@@ -87,13 +87,13 @@ contract SphereXEngineTest is Test, CFUtils {
     }
 
     function test_addAllowedPatterns_two_patterns() public activateRule(CF) {
-        int16[2] memory allowed_cf = [int16(1), -1];
+        int256[2] memory allowed_cf = [int256(1), -1];
         uint216 allowed_cf_hash = 1;
         for (uint256 i = 0; i < allowed_cf.length; i++) {
             allowed_cf_hash = uint216(bytes27(keccak256(abi.encode(int256(allowed_cf[i]), allowed_cf_hash))));
         }
 
-        int16[2] memory allowed_cf_2 = [int16(2), -2];
+        int256[2] memory allowed_cf_2 = [int256(2), -2];
         uint216 allowed_cf_hash_2 = 1;
         for (uint256 i = 0; i < allowed_cf_2.length; i++) {
             allowed_cf_hash_2 = uint216(bytes27(keccak256(abi.encode(int256(allowed_cf_2[i]), allowed_cf_hash_2))));
@@ -116,7 +116,7 @@ contract SphereXEngineTest is Test, CFUtils {
     }
 
     function test_removeAllowedPatterns(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         uint216 allowed_cf_hash = addAllowedPattern();
 
         allowed_patterns = [allowed_cf_hash];
@@ -129,11 +129,11 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // remove two cf and check that the first one was removed
     function test_removeAllowedPatterns_check_first_pattern_removed() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         uint216 allowed_cf_hash = addAllowedPattern();
-        allowed_cf_storage = [int16(2), -2];
+        allowed_cf_storage = [int256(2), -2];
         addAllowedPattern();
-        allowed_cf_storage = [int16(3), -3];
+        allowed_cf_storage = [int256(3), -3];
         uint216 allowed_cf_hash_3 = addAllowedPattern();
 
         allowed_patterns = [allowed_cf_hash, allowed_cf_hash_3];
@@ -149,11 +149,11 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // remove two cf and check that the second one was removed
     function test_removeAllowedPatterns_check_second_pattern_removed() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         uint216 allowed_cf_hash = addAllowedPattern();
-        allowed_cf_storage = [int16(2), -2];
+        allowed_cf_storage = [int256(2), -2];
         addAllowedPattern();
-        allowed_cf_storage = [int16(3), -3];
+        allowed_cf_storage = [int256(3), -3];
         uint216 allowed_cf_hash_3 = addAllowedPattern();
 
         allowed_patterns = [allowed_cf_hash, allowed_cf_hash_3];
@@ -254,7 +254,7 @@ contract SphereXEngineTest is Test, CFUtils {
     //  ============ Call flow  ============
 
     function test_sphereXValidateInternalPre_allowed_cf() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -272,7 +272,7 @@ contract SphereXEngineTest is Test, CFUtils {
     }
 
     function test_sphereXValidatePrePost_allowed_cf() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -302,10 +302,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check success with [1,2,3,4,5,-5,-4,-3,-2,-1]
     function test_CallFlow_sanity() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[10] memory allowed_long_cf = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        int256[10] memory allowed_long_cf = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         for (uint256 i = 0; i < allowed_long_cf.length; i++) {
             sendNumberToEngine(allowed_long_cf[i]);
         }
@@ -316,7 +316,7 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,-1]
     // we check success with [1,-1], [1,-1]
     function test_CallFlow_same_flow_twice_in_a_row() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -331,12 +331,12 @@ contract SphereXEngineTest is Test, CFUtils {
     }
 
     function test_currentPattern_and_callDepth_values_in_storage() public activateRule(CF) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         sendNumberToEngine(1);
         assertEq(getCurrentPattern(), uint216(bytes27(keccak256(abi.encode(int256(1), uint256(1))))));
-        assertEq(getCurrentCallDepth(), uint16(2));
+        assertEq(getCurrentCallDepth(), uint256(2));
 
         sendNumberToEngine(-1);
         assertFlowStorageSlotsInInitialState();
@@ -346,7 +346,7 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // Check that the current pattern isnt cleared at depth == 1 like in CF
     function test_PrefixTFlow_current_pattern_not_clean_at_depth_1() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -354,16 +354,16 @@ contract SphereXEngineTest is Test, CFUtils {
         }
 
         assertEq(getCurrentPattern() != uint216(1), true);
-        assertEq(getCurrentCallDepth(), uint16(1));
+        assertEq(getCurrentCallDepth(), uint256(1));
     }
 
     // Sanity check, unlike cf the current pattern isnt being cleaned at depth==1 hence this
     // test should work and the entire [1,-1,2,-2] should be approved
     function test_PrefixTFlow_sanity() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
-        allowed_cf_storage = [int16(1), -1, 2, -2];
+        allowed_cf_storage = [int256(1), -1, 2, -2];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -374,7 +374,7 @@ contract SphereXEngineTest is Test, CFUtils {
     // Check that after the block number is incremented we check the current flow from the beginning
     // In cf this test would have failed since the pattern would have been cleaned
     function test_PrefixTFlow_same_origin_same_block_number() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -391,7 +391,7 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // Check that after the block number is incremented we check the current flow from the begning
     function test_PrefixTFlow_same_origin_different_block_number() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -407,7 +407,7 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // Check that after the a different tx.origin we check the current flow from the begining
     function test_PrefixTFlow_different_origin_same_block_number() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -424,7 +424,7 @@ contract SphereXEngineTest is Test, CFUtils {
 
     function test_activateRule1_after_Rule2() public activateRule(PREFIX_TX_FLOW) {
         spherex_engine.configureRules(CF);
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -446,7 +446,7 @@ contract SphereXEngineTest is Test, CFUtils {
 
     // Check currentBlock and currentAddress state variable change
     function test_PrefixTFlow_different_origin_different_block_number() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
@@ -472,20 +472,20 @@ contract SphereXEngineTest is Test, CFUtils {
     // Check that after we recognize a new transaction we dont allow the suffix of an approved flow
     // to pass. [1,-1], [1,-1,2,-2] approved, [2,-2] reverted!
     function test_PrefixTFlow_check_suffix() public activateRule(PREFIX_TX_FLOW) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
 
-        allowed_cf_storage = [int16(1), -1, 2, -2];
+        allowed_cf_storage = [int256(1), -1, 2, -2];
         addAllowedPattern();
 
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
             sendNumberToEngine(allowed_cf_storage[i]);
         }
 
         vm.roll(2);
 
-        allowed_cf_storage = [int16(2), -2];
+        allowed_cf_storage = [int256(2), -2];
         for (uint256 i = 0; i < allowed_cf_storage.length; i++) {
             if (i == allowed_cf_storage.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -499,10 +499,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check failure with [4,3,5,-5,-3,-4]
     function test_CallFlow_partial_elements_different_length(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[6] memory not_allowed_cf = [int16(4), 3, 5, -5, -3, -4];
+        int256[6] memory not_allowed_cf = [int256(4), 3, 5, -5, -3, -4];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -514,10 +514,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check failure with [1,2,3,1,2,3,-3,-2,-1,-3,-2,-1]
     function test_CallFlow_partial_elements_same_length(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[12] memory not_allowed_cf = [int16(1), 2, 3, 1, 2, 3, -3, -2, -1, -3, -2, -1];
+        int256[12] memory not_allowed_cf = [int256(1), 2, 3, 1, 2, 3, -3, -2, -1, -3, -2, -1];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -529,10 +529,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check failure with [1,2,4,5,-5,-4,-2,-1]
     function test_CallFlow_partial_elements_same_order_different_length(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[8] memory not_allowed_cf = [int16(1), 2, 4, 5, -5, -4, -2, -1];
+        int256[8] memory not_allowed_cf = [int256(1), 2, 4, 5, -5, -4, -2, -1];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -544,10 +544,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check failure with [5,4,3,2,1,-1,-2,-3,-4,-5]
     function test_CallFlow_same_elements_backwards_order_same_length(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[10] memory not_allowed_cf = [int16(5), 4, 3, 2, 1, -1, -2, -3, -4, -5];
+        int256[10] memory not_allowed_cf = [int256(5), 4, 3, 2, 1, -1, -2, -3, -4, -5];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -559,10 +559,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check failure with [2,1,5,3,4,-4,-3,-5,-1,-2]
     function test_CallFlow_same_elements_different_order_same_length(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[10] memory not_allowed_cf = [int16(2), 1, 5, 3, 4, -4, -3, -5, -1, -2];
+        int256[10] memory not_allowed_cf = [int256(2), 1, 5, 3, 4, -4, -3, -5, -1, -2];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -577,10 +577,10 @@ contract SphereXEngineTest is Test, CFUtils {
         public
         activateRule(rule)
     {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[10] memory not_allowed_cf = [int16(1), 2, 4, 3, 5, -5, -3, -4, -2, -1];
+        int256[10] memory not_allowed_cf = [int256(1), 2, 4, 3, 5, -5, -3, -4, -2, -1];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -592,10 +592,10 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,2,3,4,5,-5,-4,-3,-2,-1]
     // we check success with [1,-1,2,-2,3,-3,4,-4,5,-5]
     function test_CallFlow_same_elements_different_nesting(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[10] memory not_allowed_cf = [int16(1), -1, 2, -2, 3, -3, 4, -4, 5, -5];
+        int256[10] memory not_allowed_cf = [int256(1), -1, 2, -2, 3, -3, 4, -4, 5, -5];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == 1) {
                 // we expect the -1 step will revert
@@ -608,12 +608,12 @@ contract SphereXEngineTest is Test, CFUtils {
     // allowed call flow is [1,-1][2,-2]
     // we check success with [1,2,-2,-1]
     function test_CallFlow_same_elements_different_nesting_2(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), -1];
+        allowed_cf_storage = [int256(1), -1];
         addAllowedPattern();
-        allowed_cf_storage = [int16(2), -2];
+        allowed_cf_storage = [int256(2), -2];
         addAllowedPattern();
 
-        int16[4] memory not_allowed_cf = [int16(1), int16(2), -2, -1];
+        int256[4] memory not_allowed_cf = [int256(1), int256(2), -2, -1];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == not_allowed_cf.length - 1) {
                 vm.expectRevert("SphereX error: disallowed tx pattern");
@@ -628,10 +628,10 @@ contract SphereXEngineTest is Test, CFUtils {
     }
 
     function test_CFNumIsZero_in_the_middle_of_a_flow(bytes8 rule) public activateRule(rule) {
-        allowed_cf_storage = [int16(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
+        allowed_cf_storage = [int256(1), 2, 3, 4, 5, -5, -4, -3, -2, -1];
         addAllowedPattern();
 
-        int16[4] memory not_allowed_cf = [int16(1), 2, 3, 0];
+        int256[4] memory not_allowed_cf = [int256(1), 2, 3, 0];
         for (uint256 i = 0; i < not_allowed_cf.length; i++) {
             if (i == 3) {
                 // we expect the 0 step will revert
